requirements.txt
# requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
openai
python-dotenv
httpx
requests